Avoid Coverity NUL termination warning about strncpy() by using memcpy() instead.  It's probably a bit more optimal in this case anyway. [1]
The program logic leading up to the creation of the strncpy/memcpy
destination buffer is a bit hairy.  Add a call to assert() to make
it clear what is happening here and detect any potential buffer
overruns in the future.

Check a couple syscall error returns.  Ignore the EEXIST error from
link() to preserve existing behavior. [2] [3]

Reported by:	Coverity
CID:		1009659 [1], 1009349 [2], 1009350 [3]


git-svn-id: https://svn.freebsd.org/base/head@299577 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f