MFC 299310: Don't store generated firmware object files in the source directory.
Trim the leading directory of a firmware source file from the resulting
target object file name so the object file is stored in the object
directory. Previously, using 'FIRMWS= /path/to/fw.bin:fw.bin' would
store the generated 'fw.bin.fwo' file in the /path/to directory. Now
it stores it in the object directory of the kernel module being built.


git-svn-id: https://svn.freebsd.org/base/stable/8@300629 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f