Set the TLB caching properties for portals at attach time.
This was found while reworking the device tree nodes for dtsec to match the
Linux device tree.  Instead of waiting and expecting later code to call
dpaa_portal_map_registers(), do the equivalent immediately upon mapping.
Otherwise, it's possible to access the pages before that function is called, and
hang the CPU.


git-svn-id: https://svn.freebsd.org/base/head@300637 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f